renamed files / deleted FirstSessionPage
