+++ src/components/ExampleComponent/index.tsx
@@ -2,7 +2,7 @@ import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
 import { useTranslation } from 'react-i18next';
-import { ExampleForm } from './examleComponent.styles';
+import { ExampleForm } from './styles';
 
 const schema = yup
   .object()
