+++ src/main.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import { BrowserRouter } from 'react-router-dom';
-import { theme } from './styles/Theme.ts';
-import { globalStyles } from './styles/GlobalStyles.ts';
 import { Global, ThemeProvider } from '@emotion/react';
 import './translation/i18n.ts';
 import { Provider } from 'react-redux';
 import store from './redux/store.ts';
+import { globalStyles } from './styles/globalStyles.ts';
+import { theme } from './styles/theme.ts';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
