--- src/pages/FirstSessionPage/FirstSessionPage.tsx
@@ -1,5 +1,0 @@
-const FirstSessionPage = () => {
-  return <div></div>;
-};
-
-export default FirstSessionPage;
