+++ src/pages/privacyPolicyPage/index.tsx
