+++ src/pages/termsOfUsePage/index.tsx
